latest

+++ transformer.py
@@ -197,7 +197,7 @@ class Transformer(nn.Module):
 
     def forward(self, src, tgt,mask_percentage:float=0.15):
         # src_mask, tgt_mask = self.generate_mask(src, tgt)
-        mask_percentage = torch.tensor(mask_percentage,dtype=torch.float32)
+        mask_percentage = torch.tensor(mask_percentage,dtype=torch.float16)
         src_mask, tgt_mask = self.generate_random_mask(src,tgt,p=mask_percentage)
 
         src_embedded = self.dropout(self.positional_encoding(self.encoder_embedding(src)))
